Some changes got lost during merge because of the merge conflicts caused due to the refactoring that has happened yesterday. Made those missing changes in the sprint-3 branch.
Tested if v1 API works. Add host on the portal worked with the appropriate TLS Policy on the DB.
v2 create API succeeded when tls_policy "insecure" is set to true.
v2 create API with tls_policy insecure as false and with an array of certificates failed with error during host agent communication (Error: server not trusted).